+++ test/test_helper.exs
@@ -10,6 +10,6 @@ Enum.each files, fn(file) ->
   Code.require_file "helper/#{file}", __DIR__
 end
 
-Code.require_file "fennec/udp/auth_template.ex", __DIR__
+Code.require_file "mongooseice/udp/auth_template.ex", __DIR__
 
 {:ok, _pid} = Helper.PortMaster.start_link()