+++ CHANGELOG.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## v0.3.0 - 2017-05-26
 
 ### Added
-* support for building Fennec as a standalone release
-* support for building Fennec as a Docker container
+* support for building MongooseICE as a standalone release
+* support for building MongooseICE as a Docker container
 * configuration of releases via system environment variables
 
 ## v0.2.0 - 2017-05-15